updated files' imports after moving of files
