+++ src/mock_data/joinRequests.js
@@ -1,11 +1,10 @@
-
 const joinRequests = [
     {
         username: 'abdulimtiaz',
         id: 1,
         message: `Hey, I think your project looks pretty cool. Mind if I work with you?`,
         projectCreator: 'anvinhtr',
-        projectName: 'BudCode'
+        projectName: 'BudCode',
     },
 
     {
@@ -13,7 +12,7 @@ const joinRequests = [
         id: 1,
         message: `I saw your project and I have a lot of the skills you need. I think I'd be a great asset!`,
         projectCreator: 'anvinhtr',
-        projectName: 'BudCode'
+        projectName: 'BudCode',
     },
 
     {
@@ -21,7 +20,7 @@ const joinRequests = [
         id: 1,
         message: `I noticed you need a front-end developer. I'm great with front-end and UX. What do you think?`,
         projectCreator: 'anvinhtr',
-        projectName: 'BudCode'
+        projectName: 'BudCode',
     },
 
     {
@@ -29,7 +28,7 @@ const joinRequests = [
         id: 1,
         message: `PLEASE LET ME JOIN!!!!!`,
         projectCreator: 'anvinhtr',
-        projectName: 'BudCode'
+        projectName: 'BudCode',
     },
 
     {
@@ -37,7 +36,7 @@ const joinRequests = [
         id: 2,
         message: `I've made a lot of organization apps in the past, check out my profile and let me know your opinion.`,
         projectCreator: 'tanviryouhana',
-        projectName: 'Organice'
+        projectName: 'Organice',
     },
 
     {
@@ -45,7 +44,7 @@ const joinRequests = [
         id: 2,
         message: `5+ years of professional HTML and CSS experience and I love your idea.`,
         projectCreator: 'tanviryouhana',
-        projectName: 'Organice'
+        projectName: 'Organice',
     },
 
     {
@@ -53,7 +52,7 @@ const joinRequests = [
         id: 3,
         message: `Hey man, I like the idea. We should work together.`,
         projectCreator: 'abdulimtiaz',
-        projectName: 'Team Up'
+        projectName: 'Team Up',
     },
 
     {
@@ -61,7 +60,7 @@ const joinRequests = [
         id: 3,
         message: `I can help a lot with the front end and have done a couple of similar projects.`,
         projectCreator: 'abdulimtiaz',
-        projectName: 'Team Up'
+        projectName: 'Team Up',
     },
 
     {
@@ -69,7 +68,7 @@ const joinRequests = [
         id: 3,
         message: `I love your idea and would love to work with you!`,
         projectCreator: 'abdulimtiaz',
-        projectName: 'Team Up'
+        projectName: 'Team Up',
     },
 
     {
@@ -77,7 +76,7 @@ const joinRequests = [
         id: 4,
         message: `I've always loved playing racing games, I'm really interested in the mechanics behind them!`,
         projectCreator: 'saif',
-        projectName: 'Cold Wheels Racing'
+        projectName: 'Cold Wheels Racing',
     },
 
     {
@@ -85,7 +84,7 @@ const joinRequests = [
         id: 4,
         message: `I have never worked on a gaming tech project, but I am open to working with you.`,
         projectCreator: 'saif',
-        projectName: 'Cold Wheels Racing'
+        projectName: 'Cold Wheels Racing',
     },
 
     {
@@ -93,7 +92,7 @@ const joinRequests = [
         id: 4,
         message: `Please allow me to work with you.`,
         projectCreator: 'saif',
-        projectName: 'Cold Wheels Racing'
+        projectName: 'Cold Wheels Racing',
     },
 
     {
@@ -101,7 +100,7 @@ const joinRequests = [
         id: 5,
         message: `I am currently a college student so this interests me a lot!`,
         projectCreator: 'tufayelahmed',
-        projectName: 'CCNY Zero'
+        projectName: 'CCNY Zero',
     },
 
     {
@@ -109,7 +108,7 @@ const joinRequests = [
         id: 5,
         message: `I'm also a CCNY student. Let's work together!`,
         projectCreator: 'tufayelahmed',
-        projectName: 'CCNY Zero'
+        projectName: 'CCNY Zero',
     },
 
     {
@@ -117,7 +116,7 @@ const joinRequests = [
         id: 6,
         message: `Using computer vision for security sounds so cool!`,
         projectCreator: 'johndoe',
-        projectName: 'Security Sentry'
+        projectName: 'Security Sentry',
     },
 
     {
@@ -125,7 +124,7 @@ const joinRequests = [
         id: 6,
         message: `I've never done computer vision but I look forward to learning a lot.`,
         projectCreator: 'johndoe',
-        projectName: 'Security Sentry'
+        projectName: 'Security Sentry',
     },
 
     {
@@ -133,7 +132,7 @@ const joinRequests = [
         id: 7,
         message: `I love working out, so this should be really useful.`,
         projectCreator: 'refatmonjur',
-        projectName: 'FitnessApp'
+        projectName: 'FitnessApp',
     },
 
     {
@@ -141,7 +140,7 @@ const joinRequests = [
         id: 7,
         message: `All the fitness apps I've used are really bloated, so I would like to collaborate on a simple one.`,
         projectCreator: 'refatmonjur',
-        projectName: 'FitnessApp'
+        projectName: 'FitnessApp',
     },
 
     {
@@ -149,7 +148,7 @@ const joinRequests = [
         id: 8,
         message: `Memes are my middle name.`,
         projectCreator: 'akbarhaider',
-        projectName: 'MemeIt'
+        projectName: 'MemeIt',
     },
 
     {
@@ -157,7 +156,7 @@ const joinRequests = [
         id: 8,
         message: `I'm not very familiar with the memes kids are using these days, but I believe it would be a "bruh moment" if you allowed me to collaborate with you.`,
         projectCreator: 'akbarhaider',
-        projectName: 'MemeIt'
+        projectName: 'MemeIt',
     },
 
     {
@@ -165,7 +164,7 @@ const joinRequests = [
         id: 9,
         message: `This seems like a challenging project, and I love a good challenge.`,
         projectCreator: 'tanzilbaraskar',
-        projectName: 'Flixter'
+        projectName: 'Flixter',
     },
 
     {
@@ -173,7 +172,7 @@ const joinRequests = [
         id: 9,
         message: `I'm what you could call a movie junkie, so I think I'd have a lot of good input.`,
         projectCreator: 'tanzilbaraskar',
-        projectName: 'Flixter'
+        projectName: 'Flixter',
     },
 
     {
@@ -181,7 +180,7 @@ const joinRequests = [
         id: 10,
         message: `I worked as a financial advisor for years prior to switching careers, so I could help a lot.`,
         projectCreator: 'mdrislam30',
-        projectName: 'Quantum Quacks'
+        projectName: 'Quantum Quacks',
     },
 
     {
@@ -189,7 +188,7 @@ const joinRequests = [
         id: 10,
         message: `I just finished up an internship at Fidelity, so I know a good amount about stocks.`,
         projectCreator: 'mdrislam30',
-        projectName: 'Quantum Quacks'
+        projectName: 'Quantum Quacks',
     },
 
     {
@@ -197,8 +196,8 @@ const joinRequests = [
         id: 10,
         message: `I'm a web app expert and I love finance. Sounds good?`,
         projectCreator: 'mdrislam30',
-        projectName: 'Quantum Quacks'
-    }
+        projectName: 'Quantum Quacks',
+    },
 ];
 
 module.exports = joinRequests;
